+++ workflow/tests/util.rb
@@ -108,7 +108,8 @@ def _run_workflow(xml, test_name, timeseries_frequency: 'none', component_loads:
         csvs[:esrat_timeseries_results] = File.join(rundir, 'results', "ESRated_ERIRatedHome_#{timeseries_frequency.capitalize}.csv")
         csvs[:esrd_timeseries_results] = File.join(rundir, 'results', "ESReference_ERIReferenceHome_#{timeseries_frequency.capitalize}.csv")
       end
-    elsif not zerh_version.nil?
+    end
+    if not zerh_version.nil?
       # Zero Energy Ready Home
       hpxmls[:zerh_ref] = File.join(rundir, 'results', 'ZERHReference.xml')
       hpxmls[:zerh_rated] = File.join(rundir, 'results', 'ZERHRated.xml')
@@ -237,12 +238,14 @@ def _get_csv_results(csvs)
       value = 1 if value == 'PASS'
       value = 0 if value == 'FAIL'
 
-      if csv.include? 'IECC'
+      if csv.include? 'IECC_'
         key = "IECC #{key}"
+      elsif csv.include? 'ES_'
+        key = "ES #{key}"
+      elsif csv.include? 'ZERH_'
+        key = "ZERH #{key}"
       end
-      if key == 'ENERGY STAR Certification' # String outputs
-        results[key] = value
-      elsif value.include? ',' # Sum values for visualization on CI
+      if value.to_s.include? ',' # Sum values for visualization on CI
         results[key] = value.split(',').map(&:to_f).sum
       else
         results[key] = Float(value)